+++ CPU/ST/CPUSingleThreadLZSS.h
@@ -2,10 +2,10 @@
 
 class CPUSingleThreadLZSS final : public AbstractLZSS {
 public:
-    std::pair<bool, double> compress(const uint8_t* inBuf, int inSize,
-        uint8_t* outBuf, int& outSize,
-        CompressFlagBlock* flagOut, int nFlagBlocks, int& flagSize) override;
+    std::pair<bool, double> compress(const uint8_t* inBuf, int64_t inSize,
+        uint8_t* outBuf, int64_t& outSize,
+        CompressFlagBlock* flagOut, int64_t nFlagBlocks, int64_t& flagSize) override;
 
-    std::pair<bool, double> decompress(CompressFlagBlock* flagIn, int nFlagBlocks,
-        const uint8_t* inBuf, int inSize, uint8_t* outBuf, int outSize) override;
+    std::pair<bool, double> decompress(CompressFlagBlock* flagIn, int64_t nFlagBlocks,
+        const uint8_t* inBuf, int64_t inSize, uint8_t* outBuf, int64_t outSize) override;
 };
